prometheusadapter: tweak some interval windows
This is to reduce flakiness. By default, our Promethesus instance will
scrape at 30s interval for all metrics. As a result, we tweak the
following windows for prometheus adapter:

1. Set `--metrics-relist-interval` to 60s. As suggested in the document,
This is the interval at which to update the cache of available metrics
from Prometheus. Since the adapter only lists metrics during discovery
that exist between the current time and the last discovery query, your
relist interval should be equal to or larger than your Prometheus scrape
interval, otherwise your metrics will occaisonally disappear from the
adapter.

2. Set `rate()` interval to be 2m. `rate()` will only be meaningful if
there are at least two samples within the interval window. It is
recommended to use a 2m interval for a scrape interval of 30s.