chore: add PR template
This template is a copy from Konvoy repo. Not only does it help
standardize PR descriptions but the underlying intent is to make sure
that we have a consistent setup for release notes. Engineers are
encouraged to add release notes whenever there is a public facing
change.

This simplifies the release process whereby now release-notes
can be generated automatically vs a dev having to manually scour PRs and
commitlog to understand what changes were made from the last release.