velero: use the new velero chart
This changed the way we deploy Minio bundled with Velero. Instead of
using Minio Operator and install a Minio instance in init container, we
switched to use Minio helm chart. We also choose to expose Minio through
Traefik ingress (on a separate entry point), instead of create another
Service type LB and handling certificates manually in the init
container.