Velero refactor (#174)
* velero: copy the config from the old release

Prepare for the update.

* velero: use the new velero chart

This changed the way we deploy Minio bundled with Velero. Instead of
using Minio Operator and install a Minio instance in init container, we
switched to use Minio helm chart. We also choose to expose Minio through
Traefik ingress (on a separate entry point), instead of create another
Service type LB and handling certificates manually in the init
container.

* traefik: copy config from previous revision

To prepare for update.

* traefik: add extra entrypoint for velero

Expose the velero minio using traefik through an extra entrypoint on
port 9000.

* test: update test group backup

Added traefik as velero now depends on traefik.

* velero: add upgrade strategy

Upgrade from pre 3.0.0 verisons require a delete and re-install. The
data won't get lost because PVC will not be deleted, and the subsequent
install will pick up those PVCs.