Add monitoring tests (#211)
* Override prometheus config

This overrides prometheus config to make it deployable on a kind
cluster. It removes the dependency on the etcd-certs secret in the
kubeaddons namespace, which is present on Konvoy. It also removes the
storageSpec to remove a dependency on persistent volumes.

* Add monitoring tests

This adds tests for prometheus, alertmanager, and grafana that verify
they serve requests.

* Factor out portForwardPodWithPrefix()