traefik: bump to chart 1.72.19
The chart change improves the charts by allowing extra SSL entrypoints
being defined, and inherit SSL configurations from global SSL config. We
observed that in some customer environment, they are using advanced
configurations for their ingress, including turning on proxy protocol on
the ingress.

Previously, we use `extraConfigEntrypoints` to add the additional velero
minio entrypoints for the velero addon. This basically means if they
have any customization on the ingress entrypoint, we'll have to ask them
to manually specifying those configs `extraConfigEntrypoints` again for
the velero-minio entrypoint, which creates a very bad user experience.

This chart change improves this experiecne by introducing
`extraSSLEntrypoints` that will inherit global configurations so that
the user does not have to do anything on the additional entrypoints.

This patch update the Addon config accordingly to use the new
`extraSSLEntrypoints`.