Fix kibana dashboard import (#511)
* Bump kibana addon revision

* Set empty API basepath for dashboard import script

The default basepath is `/`, however this causes the dashboard import
script to make requests to e.g. `//api/status`, which returns a 404.
Removing the extra `/` by emptying the basepath fixes this problem.

* Set dashboard import timeout to 180

This extends the timeout for the dashboard import script to wait for the
API to become available from 60 to 180 seconds.

Co-authored-by: Sam Tran <stran@mesosphere.com>
(cherry picked from commit c52c38ee413c9379503d00cdc671b5738fc031a8)