fix: test wrapper did only work on master branch
When trying to merge PRs into another branch (e.g. release/3.2), the
test-wrapper would still check the master branch. Now the script is
configurable using command-line arguments:

```
go run -tags experimental scripts/test-wrapper origin release/3.2
```

The first parameter is the remote's name and the second parameter is
the target branch name.

In addition to merging this commit, a change to this TeamCity build
configuration is necessary to pass these two parameters to the script:

https://teamcity.mesosphere.io/buildConfiguration/kubeaddons_KubernetesBaseAddons?mode=builds